Using emojis as pseudo-constants for the picks
The silliest approach to "legibility"! Constants are a nice way to
reduce ambiguity; emojis are cool. Unfortunately, emojis cannot be used
for constants (because they're technically lowercase). For no problem
can resist stupid meta-programming (or similar code acrobatics)! Now
everything is clearer – as long as you remember to always include a
certain module…