Strategy.parse_round_strategy follows the new encryption
Having the game logic (which pick beats which other pick) in the Pick
object (as a comparison method!) was very cute and allowed for a nice
API, but it wrapped said logic so much that exposing it in order to
get the pick that would beat other pick got really tricky.

Instead of having a central game-like object that would determine either
the outcome of two picks, or give the relevant player pick for a given
outcome and opponent pick, when need to have each Pick object exposing
being able to answer such questions as "what do you beat?" and "what
beats you?"

Plus, using emojis as pseudo-constants doesn't work anymore, since
new Pick objects must be returned – so we need to override the way picks
are converted to Hash keys if we want _any_ pick of the given value to
be used for the score computation.

In short: this implementation is completely insane. I love it. It only
needs some clean-up.