Custom object to handle the 1-index API
We want to get the stacks like with a 1-indexed array, but using an empty
first value leads to bad surprised when the raw list of stacks is accessed
(for example, to iterate over). So let's bring out the big guns and use a
custom object instead to facade everything.