Fix a nasty bug related to objects copies
OK, this one was tricky… My second solution was wrong, because the program
was operating the two cranes one after the other – but operating the first
actually altered the stacks of the second!

That was because the cranes were passed each the same array of Stack objects.
But to work on copies instead, I have to implement Stack#clone, because otherwise
only shallow copies will be made – all pointing to the same parsed data in the end.